sdr_assistant/requirements.txt
flask==2.0.1
python-dotenv==0.19.0
requests==2.26.0
pyairtable==1.4.0
Werkzeug==2.0.1
